+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/FlowState.java
@@ -1,0 +1,7 @@
+package ch.admin.bit.jeap.deploymentlog.domain;
+
+public enum FlowState {
+    OPEN,
+    CLOSED,
+    ABORTED
+}