+++ jeap-deploymentlog-domain/src/main/java/ch/admin/bit/jeap/deploymentlog/domain/FlowType.java
@@ -1,0 +1,8 @@
+package ch.admin.bit.jeap.deploymentlog.domain;
+
+public enum FlowType {
+    NEW,
+    RETRY,
+    ROLLBACK,
+    AD_HOC
+}